+++ functions.php
@@ -17,3 +17,4 @@ require_once VIRTURA_CHILD_THEME_PATH . '/inc/enqueue.php';
 require_once VIRTURA_CHILD_THEME_PATH . '/inc/carbon-fields.php';
 require_once VIRTURA_CHILD_THEME_PATH . '/inc/media-optimization.php';
 require_once VIRTURA_CHILD_THEME_PATH . '/inc/video-optimization.php';
+require_once VIRTURA_CHILD_THEME_PATH . '/inc/accessibility-images.php';